fixed issues with import and some variables that were lame, it should work now

+++ eval.py
@@ -1,18 +1,21 @@
+from os.path import isfile
+
 import numpy as np
 from vectorial_model import processing_vectorial_query
 import pickle
 import matplotlib.pyplot as plt
+from tqdm import tqdm
 
 
-def mean_average_precision(QUERIES, inverted_index, model, query_output, K):
+def mean_average_precision(QUERIES, index, model, true_results, K, weighting_schemes, stats_collection,forcebuild=False):
     average_precision = []
-    for query in QUERIES.values():
-        average_precision.append(mean_precision(query, inverted_index, model, query_output, K))
+    for i,query in QUERIES.items():
+        average_precision.append(mean_precision(query, index, model, true_results[str(i)], K, weighting_schemes, stats_collection,forcebuild))
     return np.mean(average_precision)
 
 
-def mean_precision(query, index, model, query_output, K, weighting_schemes, stats_collection):
-    results = draw_recall_precision_curve(query, index, model, query_output, weighting_schemes, stats_collection)
+def mean_precision(query, index, model, true_result, K, weighting_schemes, stats_collection,forcebuild=False):
+    results = draw_recall_precision_curve(query, index, model, true_result, weighting_schemes, stats_collection,forcebuild)
     means = np.zeros(K)
     counter = 0
     for value in np.arange(0, 1, 1 / K):
@@ -21,24 +24,28 @@ def mean_precision(query, index, model, query_output, K, weighting_schemes, stat
     return np.mean(means)
 
 
-def draw_recall_precision_curve(query, index, model, query_output, weighting_schemes, stats_collection):
+def draw_recall_precision_curve(query, index, model, true_result, weighting_schemes, stats_collection, forcebuild=False):
     weighting_scheme_query = weighting_schemes["frequency"]
     weighting_scheme_document = weighting_schemes[model]
-    processed = processing_vectorial_query(query, index, stats_collection, weighting_scheme_document,
+    if not forcebuild and isfile(model):
+        filehandler = open(model, 'rb')
+        precision_recall = pickle.load(filehandler)
+    else:
+        processed = processing_vectorial_query(query, index, stats_collection, weighting_scheme_document,
                                            weighting_scheme_query)
-    relevant_docs = list(processed.keys())
-    precision_recall = list(map(lambda x: evaluate(x, query_output, relevant_docs), tqdm(range(1, len(relevant_docs)))))
-    filehander = open(model, 'wb')
-    pickle.dump(precision_recall, filehander)
+        relevant_docs = list(processed.keys())
+        precision_recall = list(map(lambda x: evaluate(x, true_result, relevant_docs), tqdm(range(1, len(relevant_docs)))))
+        filehander = open(model, 'wb')
+        pickle.dump(precision_recall, filehander)
     return np.array(precision_recall)
 
 
-def evaluate(doc_id, query_output, ordered_relevant_docs):
-    true = set(query_output[:doc_id])
+def evaluate(doc_id, true_result, ordered_relevant_docs):
+    true = set(true_result[:doc_id])
     obtained = set(ordered_relevant_docs[:doc_id])
     false_positives = obtained - true
     true_positives = obtained.intersection(true)
-    return precision(true_positives, false_positives), recall(true_positives, query_output)
+    return precision(true_positives, false_positives), recall(true_positives, true_result)
 
 
 def precision(true_positives, false_positives):
